person-api/requirements.txt
# Person API - FastAPI

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pydantic>=2.0.0
redis>=5.0.0
